refactor: consolidate task display acceptance tests
