+++ src/test/Bau.Test.Acceptance/Bau.Test.Acceptance.csproj
@@ -81,10 +81,7 @@
     <Compile Include="DefaultTask.cs" />
     <Compile Include="Support\Baufile.cs" />
     <Compile Include="Support\MethodBaseExtensions.cs" />
-    <Compile Include="TaskListDescriptive.cs" />
-    <Compile Include="TaskListAll.cs" />
-    <Compile Include="TaskListJson.cs" />
-    <Compile Include="TaskListPrerequisites.cs" />
+    <Compile Include="DisplayingTasks.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\Bau.Test.Acceptance.CreateFile\Bau.Test.Acceptance.CreateFile.csproj">
